+++ soxspipe/commonutils/__init__.py
@@ -4,3 +4,4 @@
 from .set_of_files import set_of_files
 from .keyword_lookup import keyword_lookup
 from .detector_lookup import detector_lookup
+from .create_dispersion_map import create_dispersion_map