+++ soxspipe/recipes/_base_recipe_.py
@@ -70,7 +70,7 @@ class _base_recipe_(object):
 
         return None
 
-    def prepare_single_frame(
+    def _prepare_single_frame(
             self,
             frame,
             save=False):
@@ -82,15 +82,13 @@ class _base_recipe_(object):
         **Return:**
             - ``frame`` -- the prepared frame with mask and uncertainty extensions (CCDData object)
 
-        ---
-
         ```eval_rst
         .. todo::
 
             - write a command-line tool for this method
         ```
         """
-        self.log.debug('starting the ``prepare_single_frame`` method')
+        self.log.debug('starting the ``_prepare_single_frame`` method')
 
         kw = self.kw
         dp = self.detectorParams
@@ -109,7 +107,7 @@ class _base_recipe_(object):
         # CHECK THE NUMBER OF EXTENSIONS IS ONLY 1 AND "SOXSPIPE PRE" DOES NOT
         # EXIST. i.e. THIS IS A RAW UNTOUCHED FRAME
         if len(frame.to_hdu()) > 1 or "SOXSPIPE PRE" in frame.header:
-            raise TypeError("%(filepath)s is not a raw frame" % locals())
+            return filepath
 
         # MANIPULATE XSH DATA
         frame = self.xsh2soxs(frame)
@@ -133,14 +131,14 @@ class _base_recipe_(object):
         # NOTE FLAGS NOTE YET SUPPORTED BY CCDPROC THIS THIS WON'T GET SAVED OUT
         # AS AN EXTENSION
         arm = self.arm
-        if "NIR" in arm:
-            bitMapPath = self.calibrationRootPath + \
-                "/cal/BP_MAP_RP_%(arm)s.fits" % locals()
-        else:
+        if kw('WIN_BINX') in frame.header:
             binx = int(frame.header[kw('WIN_BINX')])
             biny = int(frame.header[kw('WIN_BINY')])
-            bitMapPath = self.calibrationRootPath + \
-                "/cal/BP_MAP_RP_%(arm)s_%(binx)sx%(biny)s.fits" % locals()
+        else:
+            binx = 1
+            biny = 1
+
+        bitMapPath = self.calibrationRootPath + "/" + dp["bad-pixel map"][f"{binx}x{biny}"]
 
         if not os.path.exists(bitMapPath):
             message = "the path to the bitMapPath %s does not exist on this machine" % (
@@ -196,9 +194,9 @@ class _base_recipe_(object):
             filenameNoExtension + "_pre" + extension
 
         # SAVE TO DISK
-        self.write(frame, filePath)
+        self._write(frame, filePath)
 
-        self.log.debug('completed the ``prepare_single_frame`` method')
+        self.log.debug('completed the ``_prepare_single_frame`` method')
         return filePath
 
     def _absolute_path(
@@ -231,9 +229,7 @@ class _base_recipe_(object):
     def prepare_frames(
             self,
             save=False):
-        """*prepare all frames in the input data*
-
-        See `prepare_single_frame` for details.
+        """*prepare raw frames by converting pixel data from ADU to electrons and adding mask and uncertainty extensions*
 
         **Key Arguments:**
             - ``save`` -- save out the prepared frame to the intermediate products directory. Default False.
@@ -249,46 +245,6 @@ class _base_recipe_(object):
         self.inputFrames = self.prepare_frames(
             save=self.settings["save-intermediate-products"])
         ```
-
-        # Preparing the Raw SOXS Frames
-
-        Here's the typical workflow for *preparing* the raw SOXS frames for data reduction:
-
-        [![](https://live.staticflickr.com/65535/50237266672_4453a01233_b.png)](https://live.staticflickr.com/65535/50237266672_4453a01233_o.png)
-
-
-        **1. Trim Overscan**
-
-        The first thing we need to do is trim off the overscan area of the image. The science-pixel regions for the detectors are read from the `soxs_detector_parameters.yaml` settings file.
-
-        **2. ADU to Electrons**
-
-        Next the pixel data is converted from ADU to electron counts by multiplying each pixel value in the raw frame by the detector gain (the gain is read in units of electrons/ADU).
-
-        $$\rm{electron\ count} = \rm{adu\ count} \times \rm{gain}$$
-
-        *3. Generating an Uncertainty Map**
-
-        Next an uncertainty map is generated for the raw image and added as the 'ERR' extension of the image.
-
-        <!-- For each pixel the uncertainty is calculated as:
-
-        $$\rm{error} = \sqrt{\rm{readnoise}^2+\rm{electron\ count}}$$ -->
-
-        <!-- **Bitmap Extension**
-
-        The appropriate bitmap extension is selected and simply added as the 'FLAG' extension of the frame. -->
-
-        **4. Bad Pixel Mask**
-
-        The default detector bitmap is read from the static calibration suite and converted to a boolean mask, with values >0 becoming TRUE to indicate these pixels need to be masks. All other values are set to FALSE. This map is add as the 'QUAL' extesion of the image. 
-
-        Finally the prepared frames are saved out into the intermediate frames location with the prefix `pre_`.
-
-        Viewing the image in DS9 (using the command `ds9 -multiframe -tile columns pre_filename.fits` to show all extensions as tiled frames) we can see the 'FLUX', 'QUAL' and 'ERR' extensions are now all present.
-
-        [![](https://live.staticflickr.com/65535/50237008782_5bb148baaf_b.png)](https://live.staticflickr.com/65535/50237008782_5bb148baaf_o.png)
-
         """
         self.log.debug('starting the ``prepare_frames`` method')
 
@@ -299,7 +255,7 @@ class _base_recipe_(object):
         frameCount = len(filepaths)
         print("# PREPARING %(frameCount)s RAW FRAMES - TRIMMING OVERSCAN, CONVERTING TO ELECTRON COUNTS, GENERATING UNCERTAINTY MAPS AND APPENDING DEFAULT BAD-PIXEL MASK" % locals())
         preframes = []
-        preframes[:] = [self.prepare_single_frame(
+        preframes[:] = [self._prepare_single_frame(
             frame=frame, save=save) for frame in filepaths]
         sof = set_of_files(
             log=self.log,
@@ -493,7 +449,7 @@ class _base_recipe_(object):
         self.log.debug('completed the ``_trim_frame`` method')
         return trimmed_frame
 
-    def write(
+    def _write(
             self,
             frame,
             filepath,
@@ -510,7 +466,7 @@ class _base_recipe_(object):
         Use within a recipe like so:
 
         ```python
-        self.write(frame, filePath)
+        self._write(frame, filePath)
         ```
         """
         self.log.debug('starting the ``write`` method')
@@ -536,20 +492,6 @@ class _base_recipe_(object):
 
         **Return:**
             - ``combined_frame`` -- the combined master frame (with updated bad-pixel and uncertainty maps)
-
-        [![](https://live.staticflickr.com/65535/50237449982_7297f60f71_b.png)](https://live.staticflickr.com/65535/50237449982_7297f60f71_o.png)
-
-        Before combining the frames we want to 'clip' any outyling pixel values found in the individual frames that are to be stacked. We isolate and remove pixels from any averaging calculation (mean or median) that have a value that strays too far from the 'typical' pixel value.
-
-        Using the median pixel value as the 'typical' value and the *median absolute deviation* (MAD) as a proxy for the standard-deviation we can accurately identify rogue pixels. For any given set of pixel values:
-
-        $$
-        MAD = \\frac{1}{N}\\sum_{i=0}^N |x_i - \\text{median}(x)|.
-        $$
-
-        The clipping is done iteratively so newly found rogue pixels are masks, median values are recaluated and clipping in repeated. The iterative process stops whenever either no more bad-pixels are to be found or the maximum number of iterations as read from the recipe settings has been reached.
-
-        After the clipping has been completed the individual frames are mean-combined, ignoring pixels in the individual bad-pixel masks. If a pixel is flagged in all individual masks it is added to the combined frame bad-pixel mask.
 
         **Usage:**
 
@@ -628,7 +570,12 @@ class _base_recipe_(object):
 
         # MASSIVE FUDGE - NEED TO CORRECTLY WRITE THE HEADER FOR COMBINED
         # IMAGES
+
         combined_frame.header = ccds[0].header
+        try:
+            combined_frame.wcs = ccds[0].wcs
+        except:
+            pass
         combined_frame.header[
             kw("DPR_CATG")] = "MASTER_%(imageType)s_%(arm)s" % locals()
 
@@ -639,6 +586,76 @@ class _base_recipe_(object):
 
         self.log.debug('completed the ``clip_and_stack`` method')
         return combined_frame
+
+    def subtract_calibrations(
+            self,
+            inputFrame,
+            master_bias=False,
+            dark=False):
+        """*subtract calibration frames from an input frame*
+
+        **Key Arguments:**
+            - ``inputFrame`` -- the input frame to have calibrations subtracted. CCDData object.
+            - ``master_bias`` -- the master bias frame to be subtracted. CCDData object. Default *False*.
+            - ``dark`` -- a dark frame to be subtracted. CCDData object. Default *False*.
+
+        **Return:**
+            - ``calibration_subtracted_frame`` -- the input frame with the calibration frame(s) subtracted. CCDData object.
+
+        **Usage:**
+
+        Within a soxspipe recipe use `subtract_calibrations` like so:
+
+        ```python
+        myCalibratedFrame = self.subtract_calibrations(
+            inputFrame=inputFrameCCDObject, master_bias=masterBiasCCDObject, dark=darkCCDObject)
+        ```
+
+        ---
+
+        ```eval_rst
+        .. todo::
+
+            - code needs written to scale dark frame to exposure time of science/calibration frame
+        ```
+        """
+        self.log.debug('starting the ``subtract_calibrations`` method')
+
+        arm = self.arm
+        kw = self.kw
+        dp = self.detectorParams
+
+        # VERIFY DATA IS IN ORDER
+        if master_bias == False and dark == False:
+            raise TypeError(
+                "subtract_calibrations method needs a master-bias frame and/or a dark frame to subtract")
+        if master_bias == False and dark.header[kw("EXPTIME")] != inputFrame.header[kw("EXPTIME")]:
+            raise AttributeError(
+                "Dark and science/calibration frame have differing exposure-times. A master-bias frame needs to be supplied to scale the dark frame to same exposure time as input science/calibration frame")
+        if master_bias != False and dark != False and dark.header[kw("EXPTIME")] != inputFrame.header[kw("EXPTIME")]:
+            raise AttributeError(
+                "CODE NEEDS WRITTEN HERE TO SCALE DARK FRAME TO EXPOSURE TIME OF SCIENCE/CALIBRATION FRAME")
+
+        # DARK WITH MATCHING EXPOSURE TIME
+        if dark != False and dark.header[kw("EXPTIME")] == inputFrame.header[kw("EXPTIME")]:
+            calibration_subtracted_frame = inputFrame.subtract(dark)
+            calibration_subtracted_frame.header = inputFrame.header
+            try:
+                calibration_subtracted_frame.wcs = inputFrame.wcs
+            except:
+                pass
+
+        # ONLY A MASTER BIAS FRAME, NO DARK
+        if dark == False and master_bias != False:
+            calibration_subtracted_frame = inputFrame.subtract(master_bias)
+            calibration_subtracted_frame.header = inputFrame.header
+            try:
+                calibration_subtracted_frame.wcs = inputFrame.wcs
+            except:
+                pass
+
+        self.log.debug('completed the ``subtract_calibrations`` method')
+        return calibration_subtracted_frame
 
     # use the tab-trigger below for new method
     # xt-class-method