+++ soxspipe/recipes/soxs_mdark.py
@@ -33,18 +33,23 @@ class soxs_mdark(_base_recipe_):
         - ``settings`` -- the settings dictionary
         - ``inputFrames`` -- input fits frames. Can be a directory, a set-of-files (SOF) file or a list of fits frame paths.
 
-    [![](https://live.staticflickr.com/65535/50266964567_2b06dfb8b8_z.png)](https://live.staticflickr.com/65535/50266964567_2b06dfb8b8_o.png)    
+    **Usage**
 
-    See `produce_product` method for usage.
+    ```python
+    from soxspipe.recipes import soxs_mdark
+    mdarkFrame = soxs_mdark(
+        log=log,
+        settings=settings,
+        inputFrames=fileList
+    )..produce_product()
+    ```
 
     ```eval_rst
     .. todo::
 
         - add a tutorial about ``soxs_mdark`` to documentation
-
     ```
     """
-    # Initialisation
 
     def __init__(
             self,
@@ -94,9 +99,6 @@ class soxs_mdark(_base_recipe_):
             self):
         """*verify the input frame match those required by the soxs_mdark recipe*
 
-        **Return:**
-            - ``None``
-
         If the fits files conform to required input for the recipe everything will pass silently, otherwise an exception shall be raised.
         """
         self.log.debug('starting the ``verify_input_frames`` method')
@@ -140,18 +142,6 @@ class soxs_mdark(_base_recipe_):
 
         **Return:**
             - ``productPath`` -- the path to the final product
-
-        **Usage**
-
-        ```python
-        from soxspipe.recipes import soxs_mdark
-        recipe = soxs_mdark(
-            log=log,
-            settings=settings,
-            inputFrames=fileList
-        )
-        mdarkFrame = recipe.produce_product()
-        ```
         """
         self.log.debug('starting the ``produce_product`` method')
 
@@ -168,7 +158,7 @@ class soxs_mdark(_base_recipe_):
             "/master_dark_%(arm)s_%(x)sx%(y)s.fits" % locals()
 
         # WRITE TO DISK
-        self.write(combined_bias_mean, productPath, overwrite=True)
+        self._write(combined_bias_mean, productPath, overwrite=True)
         self.clean_up()
 
         self.log.debug('completed the ``produce_product`` method')
@@ -176,6 +166,3 @@ class soxs_mdark(_base_recipe_):
 
     # use the tab-trigger below for new method
     # xt-class-method
-
-    # Override Method Attributes
-    # method-override-tmpx